9-Setup-Login-Screen-Components: Refactor. Introduce UseCase for SignInClient.
