+++ app/src/main/java/com/islaharper/dawnofandroid/data/repository/Repository.kt
@@ -1,5 +1,8 @@
 package com.islaharper.dawnofandroid.data.repository
 
+import android.content.Context
+import com.google.android.gms.auth.api.identity.Identity
+import com.google.android.gms.auth.api.identity.SignInClient
 import com.islaharper.dawnofandroid.domain.model.ApiResponse
 import com.islaharper.dawnofandroid.domain.model.ApiTokenRequest
 import com.islaharper.dawnofandroid.domain.repository.DataStoreOperations
@@ -52,5 +55,9 @@ class Repository @Inject constructor(
     // Verify token for access to authorised endpoints
     suspend fun verifyToken(request: ApiTokenRequest): Resource<ApiResponse> {
         return remoteDataSource.verifyToken(request)
+    }
+
+    fun getSignInClient(context: Context): SignInClient {
+        return Identity.getSignInClient(context)
     }
 }
