+++ app/src/main/java/com/islaharper/dawnofandroid/domain/useCases/signInClient/SignInClientUseCase.kt
@@ -1,0 +1,8 @@
+package com.islaharper.dawnofandroid.domain.useCases.signInClient
+
+import android.content.Context
+import com.google.android.gms.auth.api.identity.SignInClient
+
+interface SignInClientUseCase {
+    operator fun invoke(context: Context): SignInClient
+}
