+++ app/src/main/res/values/strings.xml
@@ -8,7 +8,9 @@
     <string name="successful_sign_in">Successful Sign In</string>
     <string name="sign_in">Sign In</string>
     <string name="google_logo">Google Logo</string>
-    <string name="sign_in_to_continue">Sign in to continue</string>
-    <string name="sign_in_description">To access this App you\nfirst need to sign in</string>
+    <string name="sign_in_welcome">Welcome</string>
+    <string name="sign_in_description">Sign in to continue</string>
+    <string name="sign_in_with_google">Sign in with Google</string>
+    <string name="please_wait">Please wait...</string>
 
 </resources>
